消除 size_t 转 int 的警告

+++ src/main.cpp
@@ -171,7 +171,7 @@ int main_serve(int argc, char* argv[]) {
 
 	int ret_val = engine(
 		words.data(),
-		words.size(),
+		(int) words.size(),
 		result.data(),
 		head,
 		tail,
